Stock Algo working!
Things are working as intended! This script pulls data from twelveApi, saves it, loads it, and runs data through a simple sma algorithm. If the algo says buy or sell, it is saved in a history file for later analysis. Stock class was made so that settings can be easily changed. Next step: Adding in api to actually buy things for real. Thinking of using unofficlal robinhood api for this.